+++ Sources/WebSockets/WebSocket.swift
@@ -50,16 +50,6 @@ public class WebSocket {
     }
 }
 
-public protocol RequestMiddleware : Core.Stream {
-    associatedtype Input = Request
-    associatedtype Output = Request
-}
-
-public protocol ResponseMiddleware : Core.Stream {
-    associatedtype Input = Response
-    associatedtype Output = Response
-}
-
 public class WebSocketMiddleware : RequestMiddleware {
     public typealias Input = Request
     public typealias Output = Request
